refactor: simplify android snapshot helper resolution
